+++ include/glow/GlColorMap.h
