+++ src/glow_util/RandomColorGenerator.cpp
